+++ src/components/CreateTopic/Table/CreateAlgorithm/CreateAlgorithmTable.tsx
